debundle/lowering: extract visitors (IdentifierRenamer + naturalizers)
BAZEL_TEST_INVOCATIONS=none: file move, no semantic change.
